MediaLibrary items use int64_t as their identifiers

+++ src/media/album_item.h
@@ -36,7 +36,7 @@ extern "C" {
 typedef struct album_item {
     LIBRARY_ITEM_COMMON
 
-    unsigned int i_id;
+    int64_t i_id;
     char* psz_name;
     char* psz_summary;
     time_t i_release_date;
