+++ src/ui/views/audio_list_song_view.c
@@ -36,10 +36,10 @@
 struct list_sys
 {
     LIST_VIEW_COMMON
-    unsigned int i_artist_id;
-    unsigned int i_album_id;
-    unsigned int i_genre_id;
-    unsigned int i_playlist_id;
+    int64_t i_artist_id;
+    int64_t i_album_id;
+    int64_t i_genre_id;
+    int64_t i_playlist_id;
 };
 
 struct list_view_item
